// apps/backend/src/model/kanel/public/VotableFeature.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { FiligranProduct } from '../../../__generated__/resolvers-types';
import type { VotingRoundId } from './VotingRound';

/** Identifier type for public.VotableFeature */
export type VotableFeatureId = string & { __brand: 'public.VotableFeature' };

/** Represents the table public.VotableFeature */
export default interface VotableFeature {
  id: VotableFeatureId;

  voting_round_id: VotingRoundId;

  title: string;

  short_description: string;

  description: string;

  product: FiligranProduct;

  labels: string[];

  image_url: string | null;

  position: number;

  active: boolean;

  created_at: Date;

  updated_at: Date | null;
}

/** Represents the initializer for the table public.VotableFeature */
export interface VotableFeatureInitializer {
  /** Default value: gen_random_uuid() */
  id?: VotableFeatureId;

  voting_round_id: VotingRoundId;

  title: string;

  short_description: string;

  description: string;

  product: FiligranProduct;

  /** Default value: '{}' */
  labels?: string[];

  image_url?: string | null;

  /** Default value: 0 */
  position?: number;

  /** Default value: true */
  active?: boolean;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date;

  updated_at?: Date | null;
}

/** Represents the mutator for the table public.VotableFeature */
export interface VotableFeatureMutator {
  id?: VotableFeatureId;

  voting_round_id?: VotingRoundId;

  title?: string;

  short_description?: string;

  description?: string;

  product?: FiligranProduct;

  labels?: string[];

  image_url?: string | null;

  position?: number;

  active?: boolean;

  created_at?: Date;

  updated_at?: Date | null;
}