// apps/backend/src/model/kanel/public/VotingRound.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { VotingRoundStatus } from '../../../__generated__/resolvers-types';
import type { UserId } from './User';

/** Identifier type for public.VotingRound */
export type VotingRoundId = string & { __brand: 'public.VotingRound' };

/** Represents the table public.VotingRound */
export default interface VotingRound {
  id: VotingRoundId;

  name: string;

  description: string | null;

  status: VotingRoundStatus;

  opened_at: Date | null;

  closed_at: Date | null;

  creator_id: UserId | null;

  created_at: Date;

  updated_at: Date | null;
}

/** Represents the initializer for the table public.VotingRound */
export interface VotingRoundInitializer {
  /** Default value: gen_random_uuid() */
  id?: VotingRoundId;

  name: string;

  description?: string | null;

  status: VotingRoundStatus;

  opened_at?: Date | null;

  closed_at?: Date | null;

  creator_id?: UserId | null;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date;

  updated_at?: Date | null;
}

/** Represents the mutator for the table public.VotingRound */
export interface VotingRoundMutator {
  id?: VotingRoundId;

  name?: string;

  description?: string | null;

  status?: VotingRoundStatus;

  opened_at?: Date | null;

  closed_at?: Date | null;

  creator_id?: UserId | null;

  created_at?: Date;

  updated_at?: Date | null;
}